fix window is undefined in ffmpeg worker error
