+++ src/utils/error/index.ts
@@ -100,14 +100,6 @@ export function getUserFacingErrorMessage(
     }
 }
 
-export function errorWithContext(originalError: Error, context: string) {
-    const errorWithContext = new Error(context);
-    errorWithContext.stack =
-        errorWithContext.stack.split('\n').slice(2, 4).join('\n') +
-        '\n' +
-        originalError.stack;
-    return errorWithContext;
-}
 export const parseSharingErrorCodes = (error) => {
     let parsedMessage = null;
     if (error?.status) {
